requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
easygui==0.98.3
ebcdic==1.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extract-msg==0.54.1
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fsspec==2025.7.0
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lark==1.1.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.1
olefile==0.47
oletools==0.60.2
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pcodedmp==1.2.6
pillow==11.3.0
preshed==3.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
red-black-tree-mod==1.22
regex==2024.11.6
requests==2.32.4
rich==14.0.0
RTFDE==0.1.2.1
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.5.3
segtok==1.5.11
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.1
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
win_unicode_console==0.5
wrapt==1.17.2
yake==0.6.0
